move rest of components belonging in site
