+++ .eslintrc.cjs
@@ -7,6 +7,9 @@ module.exports = {
 	overrides: [{ files: ['*.svelte', '*.svx'], processor: 'svelte3/svelte3' }],
 	settings: {
 		'svelte3/typescript': () => require('typescript'),
+		"svelte3/ignore-warnings": (warning) => {
+			return warning.code === 'a11y-click-events-have-key-events'
+		},
 	},
 	parserOptions: {
 		sourceType: 'module',
