+++ packages/parts/package.json
@@ -37,7 +37,6 @@
     "kitbook": "^0.0.24",
     "lz-string": "^1.4.4",
     "mdsvex": "^0.10.6",
-    "recordrtc": "^5.6.2",
     "replace-in-file": "^6.3.5",
     "satori": "^0.0.44",
     "satori-html": "^0.3.0",
