+++ packages/parts/src/lib/index.ts
@@ -1,11 +1,2 @@
-export { default as MediaStream } from './record/MediaStream.svelte';
-export { default as Recorder } from './record/Recorder.svelte';
-
 export { default as LatLngDisplay } from './maps/LatLngDisplay.svelte';
-export { default as Image } from './entries/media/Image.svelte';
-
-export { glossingLanguages } from './glosses/glossing-languages';
-export { semanticDomains } from './mappings/semantic-domains';
-export { partsOfSpeech } from './mappings/parts-of-speech';
-
 export { getTimeZoneLongitude } from './maps/utils/getTimeZoneLongitude';
