+++ packages/site/src/lib/components/entry/EditFieldModal.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
   import Modal from 'svelte-pieces/ui/Modal.svelte';
   import EditField from './EditField.svelte';
-  import type { Readable } from 'svelte/store';
-  export let t: Readable<any> = undefined;
 
   export let value = '';
   export let field: string;
@@ -12,5 +10,5 @@
 
 <Modal noscroll on:close>
   <span slot="heading">{display}</span>
-  <EditField on:close on:valueupdate {t} {value} {field} {display} {adding} />
+  <EditField on:close on:valueupdate {value} {field} {display} {adding} />
 </Modal>
