+++ packages/site/src/lib/components/modals/SemanticDomainsModal.svelte
@@ -6,7 +6,7 @@
   const close = () => dispatch('close');
 
   import MultiSelect from '$lib/components/ui/MultiSelect.svelte';
-  import { semanticDomains } from '@living-dictionaries/parts';
+  import { semanticDomains } from '$lib/mappings/semantic-domains';
   import type { IEntry } from '@living-dictionaries/types';
   import Button from 'svelte-pieces/ui/Button.svelte';
   export let entry: IEntry;
