+++ packages/site/src/lib/components/video/RecordVideo.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
   import { _ } from 'svelte-i18n';
   import Button from 'svelte-pieces/ui/Button.svelte';
-  import { MediaStream, Recorder } from '@living-dictionaries/parts';
+  import MediaStream from '../record/MediaStream.svelte';
+  import Recorder from '../record/Recorder.svelte';
   import { srcObject } from './srcObject';
   let videoBlob: Blob;
 </script>
