+++ packages/site/src/lib/mappings/parts-of-speech.ts
