+++ packages/site/src/routes/[dictionaryId]/entries/gallery/GalleryInfinite.broken.BAK
@@ -38,7 +38,7 @@
     refine({ isRefined: true });
   });
 
-  import { Image } from '@living-dictionaries/parts';
+  import Image from '$lib/components/image/Image.svelte';
   import { canEdit } from '$lib/stores';
   import type { IEntry } from '@living-dictionaries/types';
   let hitsType: IEntry;
