+++ packages/site/src/routes/[dictionaryId]/entry/[entryId]/EntryField.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
   import ShowHide from 'svelte-pieces/functions/ShowHide.svelte';
-  import type { Readable } from 'svelte/store';
-  export let t: Readable<any> = undefined;
   export let value: string,
     field: string = undefined,
     display: string,
@@ -41,8 +39,8 @@
       </div>
     </div>
     {#if show}
-      {#await import('@living-dictionaries/parts/src/lib/entries/EditFieldModal.svelte') then { default: EditFieldModal }}
-        <EditFieldModal on:valueupdate {t} {value} {field} {display} on:close={toggle} />
+      {#await import('$lib/components/entry/EditFieldModal.svelte') then { default: EditFieldModal }}
+        <EditFieldModal on:valueupdate {value} {field} {display} on:close={toggle} />
       {/await}
     {/if}
   </ShowHide>
