+++ packages/site/src/routes/create-dictionary/+page.svelte
@@ -12,7 +12,7 @@
   import EditableGlossesField from '$lib/components/settings/EditableGlossesField.svelte';
   import WhereSpoken from '$lib/components/settings/WhereSpoken.svelte';
   import EditableAlternateNames from '$lib/components/settings/EditableAlternateNames.svelte';
-  import { glossingLanguages } from '@living-dictionaries/parts';
+  import { glossingLanguages } from '$lib/glosses/glossing-languages';
   import SeoMetaTags from '$lib/components/SeoMetaTags.svelte';
 
   let modal: 'auth' = null;
